+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to
 
 - 🐛(backend) allow any type of extensions for media download #671
 - ♻️(frontend) improve table pdf rendering
+- 🐛(email) invitation emails in receivers language
 
 
 ## [2.2.0] - 2025-02-10